+++ Missionframework/Extensions/Discord_Log/cfgFunctions.hpp
@@ -3,3 +3,9 @@ class Factory_Log
     file = "Extensions\Discord_Log\Factory";
     class FactoryToDiscord {};
 };
+
+class FOB_Log
+{
+    file = "Extensions\Discord_Log\FOB";
+    class FOBToDiscord {};
+};