+++ Missionframework/CfgFunctions.hpp
@@ -98,6 +98,8 @@ class KPLIB {
         class getResistanceTier         {};
         class getUnitPositionId         {};
         class getUnitsCount             {};
+        class AHDNC_perFrame             {};
+        class AHDNC_perSecond            {};
         class hint                      {};
         class lambs_enableReinforcements{};
         class protectObject             {};