+++ Missionframework/Extensions/Enemy_QRF/cfgFunctions.hpp
@@ -1,6 +1,7 @@
 class Enemy_QRF
 {
     file = "Extensions\Enemy_QRF\functions";
+    
     class towerMonitoringPFH {};
     class sendQRFPatrol {};
     class spawnQRF {};