+++ Missionframework/Extensions/Drone_Jammer/cfgFunctions.hpp
@@ -1,0 +1,8 @@
+class Drone_Jammer
+{
+    file = "Extensions\Drone_Jammer\functions";
+
+    class droneJamEffects {};
+    class droneJammer {};
+};
+